Alter the order of attributes in Styles
In the schema of the stylesheet element
(https://schemas.liquid-technologies.com/officeopenxml/2006/?page=sml-styles_xsd.html)
the stylesheet is defined as an xsd:sequence which means that a
validating parser must consider the order of the elements in the
document against the order the elements are defined in the schema.

It seems that there is some version of MS Excel on MacOS that does use a
validating parser and will always insist on "repairing" the spreadsheet
with the order of the elements as they are, it will open the file fine
after that.  Passing the spreadsheet through e.g. LibreOffice (without
complaint,) will also fix it.

So this moves the numFts attribute as the first element of the
stylesheet which is how it is in the schema. I don't think any other
changes are required as .^attributes(:local) appears to return them in
the order they are defined in the class. The other attributes appear to
be in the defined order.